Move all wrapper types into one file
